reserved the original template with astroship
actually wouldn't it be better if i removed the old ones and download
a copy of their template somewhere else? I feel like this could getting
out of hands quickly since the amout of project files is like, huge.

lol. maybe do that when it's a real issue, for now just ignore anything
with `*.old` extension