panel focus responsive to scrolling
This closes #3.

The reason first try failed is because astro components can't be
assigned with DOM tags (class, id, etc.) so even the approach is right,
there is no way to locate it.

A simple work around is to wrap that panel by a <div> and assign the
class or id to that <div> wrapper.

This can be the solution if in the future need to apply css or js onto
a astro component/object.